refactor(settings): decompose 2,599-line Settings.tsx into sections
Settings.tsx becomes a 615-line shell (nav + section mounting + page
CSS); each concern moves to src/pages/settings/ owning its own state:
UpdatesSection, PlaybackSection, NetworkSection, EpgSection,
StatsSection, ParentalSection (incl. the PIN modal choreography),
AboutSection, plus a shared useSaveAnimation hook.

Kept in the shell because the behavior depended on page-level state:
updateChecking (in-flight check survives section switches) and the
EPG filter quartet (filters persist across navigation). EPG test
state re-seeds from epgTestService getters so background tests keep
running across mounts, exactly as before.

Zero behavior/visual change intended: localStorage keys, service
calls, translation keys, styles and modal flows moved verbatim.
