+++ src/pages/Settings.tsx
@@ -1,293 +1,34 @@
-import { useState, useEffect, useRef } from 'react';
-import { updateService } from '../services/updateService';
-import { playbackService } from '../services/playbackService';
-import type { PlaybackConfig } from '../services/playbackService';
-import { parentalService } from '../services/parentalService';
-import type { ParentalConfig } from '../services/parentalService';
-import { showUpToDateModal } from '../components/updateNotificationBus';
-import type { UpdateConfig } from '../types/update';
-import { usageStatsService, type UsageStats, type DailyStats } from '../services/usageStatsService';
-import { useLanguage, type SupportedLanguage } from '../services/languageService';
-import epgTestService, { type EpgTestResult, type EpgTestProgress } from '../services/epgTestService';
+import { useState, useEffect } from 'react';
+import { useLanguage } from '../services/languageService';
+import epgTestService from '../services/epgTestService';
+import { UpdatesSection } from './settings/UpdatesSection';
+import { PlaybackSection } from './settings/PlaybackSection';
+import { NetworkSection } from './settings/NetworkSection';
+import { EpgSection, type EpgResultsFilter, type EpgCountryFilter } from './settings/EpgSection';
+import { StatsSection } from './settings/StatsSection';
+import { ParentalSection } from './settings/ParentalSection';
+import { AboutSection } from './settings/AboutSection';
 
 
 export function Settings() {
-    const [updateConfig, setUpdateConfig] = useState<UpdateConfig>({
-        checkFrequency: 'on-open',
-        autoInstall: false,
-        lastCheck: 0
-    });
-    const [playbackConfig, setPlaybackConfig] = useState<PlaybackConfig>(playbackService.getConfig());
-    const [parentalConfig, setParentalConfig] = useState<ParentalConfig>(parentalService.getConfig());
-    const [checking, setChecking] = useState(false);
-    const [lastCheckDate, setLastCheckDate] = useState<string>('');
     const [activeSection, setActiveSection] = useState<string>('updates');
-    const [saveAnimation, setSaveAnimation] = useState<string | null>(null);
-    const [allowInvalidProviderCertificates, setAllowInvalidProviderCertificates] = useState(true);
-
-    // PIN Modal states
-    const [showPinModal, setShowPinModal] = useState(false);
-    const [pinInput, setPinInput] = useState(['', '', '', '']);
-    const [pinConfirm, setPinConfirm] = useState(['', '', '', '']);
-    const [pinStep, setPinStep] = useState<'enter' | 'confirm'>('enter');
-    const [pinError, setPinError] = useState('');
-    const [pinMode, setPinMode] = useState<'set' | 'verify'>('set'); // 'set' for new PIN, 'verify' for disabling
-
-    // Legal modals
-    const [showTermsModal, setShowTermsModal] = useState(false);
-    const [showPrivacyModal, setShowPrivacyModal] = useState(false);
-
-    // Usage stats
-    const [usageStats, setUsageStats] = useState<UsageStats | null>(null);
-    const [weeklyStats, setWeeklyStats] = useState<DailyStats[]>([]);
 
     // Language
-    const { language, setLanguage, t, languages } = useLanguage();
+    const { t } = useLanguage();
 
-    // EPG Test states - synced with global background service
-    const [testingEpg, setTestingEpg] = useState(epgTestService.isRunning);
-    const [epgTestProgress, setEpgTestProgress] = useState<EpgTestProgress>(epgTestService.progress);
-    const [epgTestResults, setEpgTestResults] = useState<EpgTestResult | null>(epgTestService.results);
-    const [lastEpgTestDate, setLastEpgTestDate] = useState<string | null>(epgTestService.lastTestDate);
-    const [epgResultsFilter, setEpgResultsFilter] = useState<'all' | 'working' | 'notWorking'>('all');
-    const [epgCountryFilter, setEpgCountryFilter] = useState<'all' | 'BR' | 'ARG' | 'US' | 'PT'>('all');
+    // Update check state - kept in the shell so an in-flight check survives section switches
+    const [updateChecking, setUpdateChecking] = useState(false);
+
+    // EPG result filters - kept in the shell so they persist when navigating between sections
+    const [epgResultsFilter, setEpgResultsFilter] = useState<EpgResultsFilter>('all');
+    const [epgCountryFilter, setEpgCountryFilter] = useState<EpgCountryFilter>('all');
     const [epgSearchTerm, setEpgSearchTerm] = useState('');
     const [epgCurrentPage, setEpgCurrentPage] = useState(1);
-    const EPG_ITEMS_PER_PAGE = 50;
-    const epgResultsRef = useRef<HTMLDivElement>(null);
 
-    type EpgCountryFilter = typeof epgCountryFilter;
-    type EpgWorkingItem = EpgTestResult['working'][number] & { type: 'working' };
-    type EpgNotWorkingItem = EpgTestResult['notWorking'][number] & { type: 'notWorking' };
-
-    // Subscribe to EPG test service state changes (runs in background even when navigating away)
+    // Keep the EPG background service translated even if the EPG section is never opened
     useEffect(() => {
         epgTestService.setTranslateFunction(t);
-        const unsubscribe = epgTestService.subscribe(() => {
-            setTestingEpg(epgTestService.isRunning);
-            setEpgTestProgress(epgTestService.progress);
-            setEpgTestResults(epgTestService.results);
-            setLastEpgTestDate(epgTestService.lastTestDate);
-        });
-        return unsubscribe;
     }, [t]);
-
-    // EPG handlers - delegate to background service
-    const handleEpgTest = (mode: 'full' | 'continue' | 'retryFailed' = 'full') => {
-        epgTestService.startTest(mode);
-    };
-    const handleEpgPause = () => epgTestService.pause();
-    const handleClearEpgCache = () => epgTestService.clearCache();
-
-    // Playlist channel counts by country
-    const [playlistChannelCounts, setPlaylistChannelCounts] = useState<{
-        total: number;
-        BR: number;
-        ARG: number;
-        US: number;
-        PT: number;
-    }>({ total: 0, BR: 0, ARG: 0, US: 0, PT: 0 });
-
-    // Load playlist channel counts when EPG section is active
-    useEffect(() => {
-        if (activeSection === 'epg') {
-            (async () => {
-                try {
-                    const result = await window.ipcRenderer.invoke('streams:get-live');
-                    if (result.success && result.data) {
-                        const channels = result.data as { name: string }[];
-                        let BR = 0, ARG = 0, US = 0, PT = 0;
-
-                        channels.forEach(ch => {
-                            const name = ch.name.toUpperCase();
-                            if (name.startsWith('USA:') || name.includes(' USA')) {
-                                US++;
-                            } else if (name.startsWith('ARG |') || name.startsWith('AR:')) {
-                                ARG++;
-                            } else if (name.startsWith('PT:') || name.startsWith('PT |')) {
-                                PT++;
-                            } else {
-                                BR++; // Default to BR
-                            }
-                        });
-
-                        setPlaylistChannelCounts({
-                            total: channels.length,
-                            BR, ARG, US, PT
-                        });
-                    }
-                } catch (e) {
-                    console.error('Failed to load playlist counts:', e);
-                }
-            })();
-        }
-    }, [activeSection]);
-
-    useEffect(() => {
-        loadUpdateConfig();
-        loadCertificateSettings();
-        // Load usage stats
-        setUsageStats(usageStatsService.getStats());
-        setWeeklyStats(usageStatsService.getWeeklyStats());
-    }, []);
-
-    const loadUpdateConfig = async () => {
-        const config = await updateService.getConfig();
-        setUpdateConfig(config);
-
-        if (config.lastCheck) {
-            const date = new Date(config.lastCheck);
-            setLastCheckDate(date.toLocaleString('pt-BR'));
-        }
-    };
-
-    const loadCertificateSettings = async () => {
-        try {
-            const result = await window.ipcRenderer.invoke('security:get-certificate-settings');
-            if (result.success && result.settings) {
-                setAllowInvalidProviderCertificates(Boolean(result.settings.allowInvalidProviderCertificates));
-            }
-        } catch (error) {
-            console.error('Failed to load certificate settings:', error);
-        }
-    };
-
-    const handleAllowInvalidProviderCertificatesChange = async (value: boolean) => {
-        setAllowInvalidProviderCertificates(value);
-        try {
-            const result = await window.ipcRenderer.invoke('security:set-allow-invalid-provider-certificates', value);
-            if (result.success && result.settings) {
-                setAllowInvalidProviderCertificates(Boolean(result.settings.allowInvalidProviderCertificates));
-            }
-            setSaveAnimation('allowInvalidProviderCertificates');
-            setTimeout(() => setSaveAnimation(null), 1500);
-        } catch (error) {
-            console.error('Failed to save certificate settings:', error);
-            setAllowInvalidProviderCertificates(prev => !prev);
-        }
-    };
-
-    const handleUpdateConfigChange = async <K extends keyof UpdateConfig>(key: K, value: UpdateConfig[K]) => {
-        const newConfig = { ...updateConfig, [key]: value };
-        setUpdateConfig(newConfig);
-        await updateService.setConfig(newConfig);
-
-        // Show save animation
-        setSaveAnimation(key);
-        setTimeout(() => setSaveAnimation(null), 1500);
-    };
-
-    const handlePlaybackConfigChange = <K extends keyof PlaybackConfig>(key: K, value: PlaybackConfig[K]) => {
-        const newConfig = { ...playbackConfig, [key]: value };
-        setPlaybackConfig(newConfig);
-        playbackService.setConfig({ [key]: value });
-
-        // Show save animation
-        setSaveAnimation(key);
-        setTimeout(() => setSaveAnimation(null), 1500);
-    };
-
-    const handleParentalConfigChange = <K extends keyof ParentalConfig>(key: K, value: ParentalConfig[K]) => {
-        // When trying to disable parental control, require PIN verification
-        if (key === 'enabled' && value === false && parentalService.hasPin()) {
-            resetPinModal();
-            setPinMode('verify');
-            setShowPinModal(true);
-            return; // Don't change config until PIN is verified
-        }
-
-        const newConfig = { ...parentalConfig, [key]: value };
-        setParentalConfig(newConfig);
-        parentalService.setConfig({ [key]: value });
-
-        // When enabling parental control for the first time, prompt for PIN
-        if (key === 'enabled' && value === true && !parentalService.hasPin()) {
-            resetPinModal();
-            setPinMode('set');
-            setShowPinModal(true);
-        }
-
-        // Show save animation
-        setSaveAnimation(`parental_${key}`);
-        setTimeout(() => setSaveAnimation(null), 1500);
-    };
-
-    const handlePinSubmit = async () => {
-        const pin = pinInput.join('');
-
-        if (pin.length !== 4) {
-            setPinError(t('parental', 'pinError4Digits'));
-            return;
-        }
-
-        // Verification mode - check if PIN is correct to disable parental control
-        if (pinMode === 'verify') {
-            if (await parentalService.verifyPin(pin)) {
-                // PIN is correct - disable parental control
-                setParentalConfig(prev => ({ ...prev, enabled: false }));
-                parentalService.setConfig({ enabled: false });
-                setShowPinModal(false);
-                resetPinModal();
-
-                // Show save animation
-                setSaveAnimation('parental_enabled');
-                setTimeout(() => setSaveAnimation(null), 1500);
-            } else {
-                setPinError(t('parental', 'pinIncorrect'));
-                setPinInput(['', '', '', '']);
-            }
-            return;
-        }
-
-        // Set mode - setting a new PIN
-        if (pinStep === 'enter') {
-            // Move to confirm step
-            setPinStep('confirm');
-            setPinError('');
-        } else {
-            // Confirm PIN
-            const confirmPin = pinConfirm.join('');
-            if (pin !== confirmPin) {
-                setPinError(t('parental', 'pinMismatch'));
-                setPinConfirm(['', '', '', '']);
-                return;
-            }
-            // Save PIN
-            await parentalService.setPin(pin);
-            setParentalConfig(parentalService.getConfig());
-            setShowPinModal(false);
-            resetPinModal();
-
-            // Show save animation
-            setSaveAnimation('parental_pin');
-            setTimeout(() => setSaveAnimation(null), 1500);
-        }
-    };
-
-    const resetPinModal = () => {
-        setPinInput(['', '', '', '']);
-        setPinConfirm(['', '', '', '']);
-        setPinStep('enter');
-        setPinError('');
-    };
-
-    const handleCheckNow = async () => {
-        setChecking(true);
-        try {
-            // This will trigger update:available event if update exists
-            // If no update, we show the "up to date" modal manually
-            const result = await updateService.checkForUpdates();
-            if (!result.updateAvailable) {
-                showUpToDateModal();
-            }
-            // If update is available, the UpdateNotification will show automatically
-            await loadUpdateConfig();
-        } catch (error) {
-            console.error('Error checking for updates:', error);
-        } finally {
-            setChecking(false);
-        }
-    };
 
     const sections = [
         { id: 'updates', icon: '🔄', label: t('nav', 'updates'), color: '#10b981' },
@@ -298,23 +39,6 @@ export function Settings() {
         { id: 'parental', icon: '👨‍👩‍👧', label: t('nav', 'parental'), color: '#ef4444' },
         { id: 'about', icon: 'ℹ️', label: t('nav', 'about'), color: '#f59e0b' }
     ];
-
-    // Helper functions for stats display
-    const formatWatchTime = (seconds: number) => {
-        const hours = Math.floor(seconds / 3600);
-        const minutes = Math.floor((seconds % 3600) / 60);
-        if (hours > 0) return `${hours}h ${minutes}min`;
-        return `${minutes}min`;
-    };
-
-    const getWeekDayLabel = (dateStr: string) => {
-        const date = new Date(dateStr + 'T12:00:00');
-        return ['D', 'S', 'T', 'Q', 'Q', 'S', 'S'][date.getDay()];
-    };
-
-    const getMaxWeeklySeconds = () => {
-        return Math.max(...weeklyStats.map(d => d.totalSeconds), 1);
-    };
 
     return (
         <>
@@ -354,1747 +78,39 @@ export function Settings() {
                     <div className="settings-content">
                         {/* Updates Section */}
                         {activeSection === 'updates' && (
-                            <div className="section-card">
-                                <div className="section-header">
-                                    <div className="section-icon" style={{ background: 'linear-gradient(135deg, #10b981, #059669)' }}>🔄</div>
-                                    <div>
-                                        <h2>{t('updates', 'title')}</h2>
-                                        <p>{t('updates', 'description')}</p>
-                                    </div>
-                                </div>
-
-                                <div className="settings-group">
-                                    {/* Check Frequency */}
-                                    <div className="setting-item">
-                                        <div className="setting-info">
-                                            <label>{t('updates', 'checkFrequency')}</label>
-                                            <p>{t('updates', 'checkFrequencyDesc')}</p>
-                                        </div>
-                                        <select
-                                            className="setting-select"
-                                            value={updateConfig.checkFrequency}
-                                            onChange={(e) => handleUpdateConfigChange('checkFrequency', e.target.value as UpdateConfig['checkFrequency'])}
-                                        >
-                                            <option value="on-open">{t('updates', 'onOpen')}</option>
-                                            <option value="1-day">{t('updates', 'daily')}</option>
-                                            <option value="1-week">{t('updates', 'weekly')}</option>
-                                            <option value="1-month">{t('updates', 'monthly')}</option>
-                                        </select>
-                                    </div>
-
-                                    {/* Auto Install Toggle */}
-                                    <div className="setting-item">
-                                        <div className="setting-info">
-                                            <label>{t('updates', 'autoInstall')}</label>
-                                            <p>{t('updates', 'autoInstallDesc')}</p>
-                                        </div>
-                                        <label className="toggle-switch">
-                                            <input
-                                                type="checkbox"
-                                                checked={updateConfig.autoInstall}
-                                                onChange={(e) => handleUpdateConfigChange('autoInstall', e.target.checked)}
-                                            />
-                                            <span className="toggle-slider"></span>
-                                        </label>
-                                        {saveAnimation === 'autoInstall' && (
-                                            <span className="save-indicator">✓ Salvo</span>
-                                        )}
-                                    </div>
-
-                                    <div className="setting-item">
-                                        <div className="setting-info">
-                                            <label>{t('updates', 'language')}</label>
-                                            <p>{t('updates', 'languageDesc')}</p>
-                                        </div>
-                                        <select
-                                            className="setting-select"
-                                            value={language}
-                                            onChange={(e) => {
-                                                setLanguage(e.target.value as SupportedLanguage);
-                                                setSaveAnimation('language');
-                                                setTimeout(() => setSaveAnimation(null), 1500);
-                                            }}
-                                        >
-                                            {languages.map(lang => (
-                                                <option key={lang.code} value={lang.code}>
-                                                    {lang.flag} {lang.name}
-                                                </option>
-                                            ))}
-                                        </select>
-                                        {saveAnimation === 'language' && (
-                                            <span className="save-indicator">{t('settings', 'saved')}</span>
-                                        )}
-                                    </div>
-
-                                    {/* Last Check */}
-                                    {lastCheckDate && (
-                                        <div className="last-check">
-                                            <span className="check-icon">🕐</span>
-                                            <span>Última verificação: <strong>{lastCheckDate}</strong></span>
-                                        </div>
-                                    )}
-
-                                    {/* Check Now Button */}
-                                    <button
-                                        className={`check-btn ${checking ? 'checking' : ''}`}
-                                        onClick={handleCheckNow}
-                                        disabled={checking}
-                                    >
-                                        {checking ? (
-                                            <>
-                                                <span className="spinner"></span>
-                                                <span>Verificando...</span>
-                                            </>
-                                        ) : (
-                                            <>
-                                                <span>🔍</span>
-                                                <span>Verificar Atualizações Agora</span>
-                                            </>
-                                        )}
-                                    </button>
-                                </div>
-                            </div>
+                            <UpdatesSection checking={updateChecking} setChecking={setUpdateChecking} />
                         )}
 
                         {/* Playback Section */}
-                        {activeSection === 'playback' && (
-                            <div className="section-card">
-                                <div className="section-header">
-                                    <div className="section-icon" style={{ background: 'linear-gradient(135deg, #3b82f6, #2563eb)' }}>⏯️</div>
-                                    <div>
-                                        <h2>{t('playback', 'title')}</h2>
-                                        <p>{t('playback', 'description')}</p>
-                                    </div>
-                                </div>
-
-                                <div className="settings-group">
-                                    <div className="setting-item">
-                                        <div className="setting-info">
-                                            <label>{t('playback', 'bufferSize')}</label>
-                                            <p>{t('playback', 'bufferSizeDesc')}</p>
-                                        </div>
-                                        <select
-                                            className="setting-select"
-                                            value={playbackConfig.bufferSize}
-                                            onChange={(e) => handlePlaybackConfigChange('bufferSize', e.target.value as PlaybackConfig['bufferSize'])}
-                                        >
-                                            <option value="intelligent">{t('playback', 'intelligent')}</option>
-                                            <option value="5">5 {t('playback', 'seconds')}</option>
-                                            <option value="10">10 {t('playback', 'seconds')}</option>
-                                            <option value="15">15 {t('playback', 'seconds')}</option>
-                                            <option value="30">30 {t('playback', 'seconds')}</option>
-                                        </select>
-                                        {saveAnimation === 'bufferSize' && <span className="save-indicator">{t('settings', 'saved')}</span>}
-                                    </div>
-
-                                    <div className="setting-item">
-                                        <div className="setting-info">
-                                            <label>{t('playback', 'videoCodec')}</label>
-                                            <p>{t('playback', 'videoCodecDesc')}</p>
-                                        </div>
-                                        <select
-                                            className="setting-select"
-                                            value={playbackConfig.videoCodec}
-                                            onChange={(e) => handlePlaybackConfigChange('videoCodec', e.target.value as PlaybackConfig['videoCodec'])}
-                                        >
-                                            <option value="auto">Auto</option>
-                                            <option value="h264">H.264 (AVC)</option>
-                                            <option value="h265">H.265 (HEVC)</option>
-                                            <option value="vp9">VP9</option>
-                                        </select>
-                                        {saveAnimation === 'videoCodec' && <span className="save-indicator">{t('settings', 'saved')}</span>}
-                                    </div>
-
-                                    {/* Subtitle Language Setting */}
-                                    <div className="setting-item">
-                                        <div className="setting-info">
-                                            <label>{t('playback', 'subtitleLanguage') || 'Idioma das Legendas'}</label>
-                                            <p>{t('playback', 'subtitleLanguageDesc') || 'Idioma preferido para download automático de legendas'}</p>
-                                        </div>
-                                        <select
-                                            className="setting-select"
-                                            value={playbackConfig.subtitleLanguage}
-                                            onChange={(e) => handlePlaybackConfigChange('subtitleLanguage', e.target.value as PlaybackConfig['subtitleLanguage'])}
-                                        >
-                                            <option value="pt-br">🇧🇷 Português (Brasil)</option>
-                                            <option value="pt">🇵🇹 Português</option>
-                                            <option value="en">🇺🇸 English</option>
-                                            <option value="es">🇪🇸 Español</option>
-                                        </select>
-                                        {saveAnimation === 'subtitleLanguage' && <span className="save-indicator">{t('settings', 'saved')}</span>}
-                                    </div>
-
-                                    {/* Forced Subtitles Setting */}
-                                    <div className="setting-item">
-                                        <div className="setting-info">
-                                            <label>{t('playback', 'forcedSubtitles') || 'Legendas Forçadas'}</label>
-                                            <p>{t('playback', 'forcedSubtitlesDesc') || 'Carregar automaticamente legendas de placas e diálogos estrangeiros (não funciona em conteúdo [L] já legendado)'}</p>
-                                        </div>
-                                        <label className="toggle-switch">
-                                            <input
-                                                type="checkbox"
-                                                checked={playbackConfig.forcedSubtitlesEnabled}
-                                                onChange={(e) => handlePlaybackConfigChange('forcedSubtitlesEnabled', e.target.checked)}
-                                            />
-                                            <span className="toggle-slider"></span>
-                                        </label>
-                                        {saveAnimation === 'forcedSubtitlesEnabled' && <span className="save-indicator">{t('settings', 'saved')}</span>}
-                                    </div>
-
-                                    {/* Auto Play Next Episode Setting */}
-                                    <div className="setting-item">
-                                        <div className="setting-info">
-                                            <label>{t('playback', 'autoPlayNext')}</label>
-                                            <p>{t('playback', 'autoPlayNextDesc')}</p>
-                                        </div>
-                                        <label className="toggle-switch">
-                                            <input
-                                                type="checkbox"
-                                                checked={playbackConfig.autoPlayNextEpisode}
-                                                onChange={(e) => handlePlaybackConfigChange('autoPlayNextEpisode', e.target.checked)}
-                                            />
-                                            <span className="toggle-slider"></span>
-                                        </label>
-                                        {saveAnimation === 'autoPlayNextEpisode' && <span className="save-indicator">{t('settings', 'saved')}</span>}
-                                    </div>
-
-                                    {/* Skip Intro Setting */}
-                                    <div className="setting-item">
-                                        <div className="setting-info">
-                                            <label>{t('playback', 'skipIntro')}</label>
-                                            <p>{t('playback', 'skipIntroDesc')}</p>
-                                        </div>
-                                        <label className="toggle-switch">
-                                            <input type="checkbox" defaultChecked />
-                                            <span className="toggle-slider"></span>
-                                        </label>
-                                    </div>
-
-                                    {/* Click-Through Mode Setting */}
-                                    <div className="setting-item">
-                                        <div className="setting-info">
-                                            <label>{t('playback', 'clickThrough')}</label>
-                                            <p>{t('playback', 'clickThroughDesc')}</p>
-                                        </div>
-                                        <label className="toggle-switch">
-                                            <input
-                                                type="checkbox"
-                                                checked={playbackConfig.clickThroughEnabled}
-                                                onChange={(e) => handlePlaybackConfigChange('clickThroughEnabled', e.target.checked)}
-                                            />
-                                            <span className="toggle-slider"></span>
-                                        </label>
-                                        {saveAnimation === 'clickThroughEnabled' && <span className="save-indicator">{t('settings', 'saved')}</span>}
-                                    </div>
-                                </div>
-                            </div>
-                        )}
+                        {activeSection === 'playback' && <PlaybackSection />}
 
                         {/* Network Section */}
-                        {activeSection === 'network' && (
-                            <div className="section-card">
-                                <div className="section-header">
-                                    <div className="section-icon" style={{ background: 'linear-gradient(135deg, #14b8a6, #0f766e)' }}>🔐</div>
-                                    <div>
-                                        <h2>Rede e certificados</h2>
-                                        <p>Controle de compatibilidade para provedores IPTV com TLS ou CORS antigo ou mal configurado.</p>
-                                    </div>
-                                </div>
-
-                                <div className="settings-group">
-                                    <div className="setting-item">
-                                        <div className="setting-info">
-                                            <label>Modo compatível com certificados inválidos</label>
-                                            <p>Permite certificados inválidos e compatibilidade CORS somente para o provedor IPTV configurado e subdomínios relacionados. Útil para provedores antigos, mas reduz a segurança dessa conexão.</p>
-                                        </div>
-                                        <label className="toggle-switch">
-                                            <input
-                                                type="checkbox"
-                                                checked={allowInvalidProviderCertificates}
-                                                onChange={(e) => handleAllowInvalidProviderCertificatesChange(e.target.checked)}
-                                            />
-                                            <span className="toggle-slider"></span>
-                                        </label>
-                                        {saveAnimation === 'allowInvalidProviderCertificates' && (
-                                            <span className="save-indicator">{t('settings', 'saved')}</span>
-                                        )}
-                                    </div>
-
-                                    {allowInvalidProviderCertificates && (
-                                        <div className="certificate-warning">
-                                            <strong>Atenção:</strong> este modo não libera certificados inválidos nem CORS para TMDB, atualizações, GitHub ou domínios externos independentes. Ele vale apenas para o servidor IPTV salvo no app e hosts relacionados aprovados.
-                                        </div>
-                                    )}
-                                </div>
-                            </div>
-                        )}
+                        {activeSection === 'network' && <NetworkSection />}
 
                         {/* EPG Section - Test runs in background */}
                         {activeSection === 'epg' && (
-                            <div className="section-card">
-                                {/* EPG Animations */}
-                                <style>{`
-                                    @keyframes epgPulse {
-                                        0%, 100% { opacity: 1; }
-                                        50% { opacity: 0.7; }
-                                    }
-                                    @keyframes epgGlow {
-                                        0%, 100% { box-shadow: 0 0 15px rgba(6, 182, 212, 0.3); }
-                                        50% { box-shadow: 0 0 25px rgba(6, 182, 212, 0.5); }
-                                    }
-                                    @keyframes epgSlideIn {
-                                        from { opacity: 0; transform: translateY(10px); }
-                                        to { opacity: 1; transform: translateY(0); }
-                                    }
-                                    @keyframes epgProgressShine {
-                                        0% { background-position: -200% 0; }
-                                        100% { background-position: 200% 0; }
-                                    }
-                                    .epg-grid-item {
-                                        transition: all 0.3s cubic-bezier(0.4, 0, 0.2, 1);
-                                        cursor: default;
-                                    }
-                                    .epg-grid-item:hover {
-                                        transform: translateY(-3px) scale(1.02);
-                                        box-shadow: 0 8px 20px rgba(0, 0, 0, 0.3);
-                                    }
-                                    .epg-country-card {
-                                        transition: all 0.3s cubic-bezier(0.4, 0, 0.2, 1);
-                                        cursor: pointer;
-                                    }
-                                    .epg-country-card:hover {
-                                        transform: translateY(-4px) scale(1.05);
-                                        box-shadow: 0 10px 25px rgba(0, 0, 0, 0.25);
-                                    }
-                                    .epg-channel-item {
-                                        transition: all 0.2s ease;
-                                        animation: epgSlideIn 0.3s ease forwards;
-                                    }
-                                    .epg-channel-item:hover {
-                                        transform: translateX(4px);
-                                        box-shadow: 0 4px 12px rgba(0, 0, 0, 0.2);
-                                    }
-                                    .epg-progress-bar {
-                                        background: linear-gradient(90deg, #06b6d4, #8b5cf6, #06b6d4);
-                                        background-size: 200% 100%;
-                                        animation: epgProgressShine 2s linear infinite;
-                                    }
-                                    .epg-testing-badge {
-                                        animation: epgPulse 1.5s ease-in-out infinite;
-                                    }
-                                    .epg-filter-btn {
-                                        transition: all 0.25s ease;
-                                    }
-                                    .epg-filter-btn:hover {
-                                        transform: translateY(-2px);
-                                    }
-                                    .epg-filter-btn:active {
-                                        transform: scale(0.95);
-                                    }
-                                    .epg-page-btn {
-                                        transition: all 0.25s cubic-bezier(0.4, 0, 0.2, 1);
-                                        position: relative;
-                                        overflow: hidden;
-                                    }
-                                    .epg-page-btn:hover:not(:disabled) {
-                                        transform: translateY(-3px) scale(1.05);
-                                        box-shadow: 0 6px 20px rgba(6, 182, 212, 0.4);
-                                    }
-                                    .epg-page-btn:active:not(:disabled) {
-                                        transform: scale(0.92);
-                                    }
-                                    .epg-page-btn:disabled {
-                                        opacity: 0.5;
-                                    }
-                                    .epg-page-indicator {
-                                        transition: all 0.3s ease;
-                                        animation: epgPulse 2s ease-in-out infinite;
-                                    }
-                                `}</style>
-                                <div className="section-header">
-                                    <div className="section-icon" style={{ background: 'linear-gradient(135deg, #06b6d4, #0891b2)' }}>📡</div>
-                                    <div>
-                                        <h2>Teste de EPG</h2>
-                                        <p>Verifique se seus canais têm guia de programação. O teste continua mesmo navegando.</p>
-                                    </div>
-                                </div>
-
-                                <div className="settings-group">
-                                    {/* Playlist Channel Counts Grid - Redesigned */}
-                                    {playlistChannelCounts.total > 0 && (
-                                        <div style={{ marginBottom: '24px' }}>
-                                            <label style={{ color: 'rgba(255,255,255,0.5)', fontSize: '11px', marginBottom: '12px', display: 'block', textTransform: 'uppercase', letterSpacing: '1px' }}>
-                                                📺 {t('epg', 'playlistChannels')}
-                                            </label>
-                                            {/* Featured Total Card */}
-                                            <div className="setting-item epg-grid-item" style={{
-                                                padding: '20px',
-                                                background: 'linear-gradient(135deg, rgba(168, 85, 247, 0.2), rgba(236, 72, 153, 0.15))',
-                                                border: '1px solid rgba(168, 85, 247, 0.4)',
-                                                marginBottom: '12px',
-                                                justifyContent: 'center'
-                                            }}>
-                                                <div style={{ display: 'flex', alignItems: 'center', justifyContent: 'center', gap: '16px' }}>
-                                                    <div style={{ fontSize: '42px', fontWeight: 800, color: '#a855f7' }}>
-                                                        {playlistChannelCounts.total}
-                                                    </div>
-                                                    <div style={{ textAlign: 'left' }}>
-                                                        <div style={{ color: 'white', fontSize: '16px', fontWeight: 600 }}>{t('epg', 'channelsLabel')}</div>
-                                                        <div style={{ color: 'rgba(255,255,255,0.5)', fontSize: '12px' }}>{t('epg', 'availableForTest')}</div>
-                                                    </div>
-                                                </div>
-                                            </div>
-                                            {/* Countries Row */}
-                                            <div style={{ display: 'grid', gridTemplateColumns: 'repeat(4, 1fr)', gap: '10px' }}>
-                                                {[
-                                                    { code: 'BR', flag: '🇧🇷', count: playlistChannelCounts.BR, color: '#22c55e', label: 'Brasil' },
-                                                    { code: 'ARG', flag: '🇦🇷', count: playlistChannelCounts.ARG, color: '#60a5fa', label: 'Argentina' },
-                                                    { code: 'US', flag: '🇺🇸', count: playlistChannelCounts.US, color: '#f59e0b', label: 'EUA' },
-                                                    { code: 'PT', flag: '🇵🇹', count: playlistChannelCounts.PT, color: '#ef4444', label: 'Portugal' }
-                                                ].map(({ code, flag, count, color, label }) => (
-                                                    <div key={code} className="setting-item epg-grid-item" style={{
-                                                        padding: '12px 8px',
-                                                        background: `rgba(${color === '#22c55e' ? '34, 197, 94' : color === '#60a5fa' ? '96, 165, 250' : color === '#f59e0b' ? '245, 158, 11' : '239, 68, 68'}, 0.1)`,
-                                                        border: `1px solid rgba(${color === '#22c55e' ? '34, 197, 94' : color === '#60a5fa' ? '96, 165, 250' : color === '#f59e0b' ? '245, 158, 11' : '239, 68, 68'}, 0.25)`,
-                                                        justifyContent: 'center'
-                                                    }}>
-                                                        <div style={{ textAlign: 'center' }}>
-                                                            <div style={{ fontSize: '18px', marginBottom: '4px' }}>{flag}</div>
-                                                            <div style={{ fontSize: '20px', fontWeight: 800, color }}>{count}</div>
-                                                            <div style={{ color: 'rgba(255,255,255,0.4)', fontSize: '10px', marginTop: '2px' }}>{label}</div>
-                                                        </div>
-                                                    </div>
-                                                ))}
-                                            </div>
-                                        </div>
-                                    )}
-
-                                    {/* Status Info */}
-                                    {lastEpgTestDate && !testingEpg && (
-                                        <div className="last-check">
-                                            <span className="check-icon">📅</span>
-                                            <span>{t('epg', 'lastTest')} <strong>{lastEpgTestDate}</strong></span>
-                                        </div>
-                                    )}
-
-                                    {/* Progress when testing */}
-                                    {testingEpg && (
-                                        <div className="setting-item" style={{
-                                            background: 'linear-gradient(135deg, rgba(6, 182, 212, 0.15), rgba(139, 92, 246, 0.1))',
-                                            border: '1px solid rgba(6, 182, 212, 0.3)'
-                                        }}>
-                                            <div style={{ flex: 1 }}>
-                                                <div style={{ display: 'flex', justifyContent: 'space-between', marginBottom: '12px' }}>
-                                                    <label className="epg-testing-badge" style={{ color: '#06b6d4', fontWeight: 700 }}>
-                                                        ⏳ {t('epg', 'testingBackground')}
-                                                    </label>
-                                                    <span style={{ color: 'rgba(255,255,255,0.7)', fontWeight: 600 }}>
-                                                        {epgTestProgress.current} / {epgTestProgress.total}
-                                                    </span>
-                                                </div>
-                                                <div style={{
-                                                    height: '8px',
-                                                    background: 'rgba(255,255,255,0.1)',
-                                                    borderRadius: '4px',
-                                                    overflow: 'hidden',
-                                                    marginBottom: '8px'
-                                                }}>
-                                                    <div className="epg-progress-bar" style={{
-                                                        width: `${epgTestProgress.total > 0 ? (epgTestProgress.current / epgTestProgress.total) * 100 : 0}%`,
-                                                        height: '100%',
-                                                        borderRadius: '4px',
-                                                        transition: 'width 0.3s ease'
-                                                    }} />
-                                                </div>
-                                                <p style={{ margin: 0, fontSize: '13px', color: 'rgba(255,255,255,0.5)' }}>
-                                                    📺 {epgTestProgress.currentChannel}
-                                                </p>
-                                            </div>
-                                        </div>
-                                    )}
-
-                                    {/* Main Action Button */}
-                                    {!testingEpg ? (
-                                        <button className="check-btn" onClick={() => handleEpgTest('full')} style={{
-                                            background: 'linear-gradient(135deg, #06b6d4 0%, #0891b2 100%)'
-                                        }}>
-                                            <span>🧪</span>
-                                            <span>{t('epg', 'testAllChannels')}</span>
-                                        </button>
-                                    ) : (
-                                        <button className="check-btn" onClick={handleEpgPause} style={{
-                                            background: 'linear-gradient(135deg, #ef4444 0%, #dc2626 100%)'
-                                        }}>
-                                            <span>⏸️</span>
-                                            <span>{t('epg', 'pauseTest')}</span>
-                                        </button>
-                                    )}
-
-                                    {/* Secondary Actions */}
-                                    {!testingEpg && epgTestResults && (
-                                        <div style={{ display: 'flex', gap: '12px', flexWrap: 'wrap' }}>
-                                            {epgTestResults.lastScannedIndex && epgTestResults.lastScannedIndex < (epgTestResults.summary?.total || 0) && (
-                                                <button className="check-btn" onClick={() => handleEpgTest('continue')} style={{
-                                                    flex: 1,
-                                                    background: 'linear-gradient(135deg, #8b5cf6 0%, #7c3aed 100%)'
-                                                }}>
-                                                    <span>▶️</span>
-                                                    <span>{t('epg', 'continueFrom')} ({epgTestResults.lastScannedIndex}/{epgTestResults.summary?.total})</span>
-                                                </button>
-                                            )}
-                                            {epgTestResults.notWorking.length > 0 && (
-                                                <button className="check-btn" onClick={() => handleEpgTest('retryFailed')} style={{
-                                                    flex: 1,
-                                                    background: 'linear-gradient(135deg, #f59e0b 0%, #d97706 100%)'
-                                                }}>
-                                                    <span>🔁</span>
-                                                    <span>{t('epg', 'retryFailed')} ({epgTestResults.notWorking.length})</span>
-                                                </button>
-                                            )}
-                                            <button className="check-btn" onClick={handleClearEpgCache} style={{
-                                                flex: 1,
-                                                background: 'rgba(255,255,255,0.1)',
-                                                color: 'rgba(255,255,255,0.7)'
-                                            }}>
-                                                <span>🗑️</span>
-                                                <span>{t('epg', 'clearCache')}</span>
-                                            </button>
-                                        </div>
-                                    )}
-                                    {/* Results Summary */}
-                                    {epgTestResults && (
-                                        <div ref={epgResultsRef} style={{ marginTop: '24px' }}>
-                                            {/* Success Rate Bar */}
-                                            <div style={{ marginBottom: '20px' }}>
-                                                <div style={{ display: 'flex', justifyContent: 'space-between', alignItems: 'center', marginBottom: '8px' }}>
-                                                    <span style={{ color: 'rgba(255,255,255,0.6)', fontSize: '12px' }}>{t('epg', 'successRate')}</span>
-                                                    <span style={{ color: '#22c55e', fontWeight: 700, fontSize: '14px' }}>
-                                                        {Math.round((epgTestResults.working.length / (epgTestResults.working.length + epgTestResults.notWorking.length || 1)) * 100)}%
-                                                    </span>
-                                                </div>
-                                                <div style={{
-                                                    height: '12px',
-                                                    background: 'rgba(239, 68, 68, 0.3)',
-                                                    borderRadius: '6px',
-                                                    overflow: 'hidden',
-                                                    display: 'flex'
-                                                }}>
-                                                    <div style={{
-                                                        width: `${(epgTestResults.working.length / (epgTestResults.working.length + epgTestResults.notWorking.length || 1)) * 100}%`,
-                                                        background: 'linear-gradient(90deg, #22c55e, #16a34a)',
-                                                        borderRadius: '6px',
-                                                        transition: 'width 0.5s ease'
-                                                    }} />
-                                                </div>
-                                            </div>
-
-                                            {/* Summary Cards - Horizontal */}
-                                            <div style={{ display: 'grid', gridTemplateColumns: '1fr 1fr', gap: '12px', marginBottom: '24px' }}>
-                                                <div className="setting-item epg-grid-item" style={{
-                                                    background: 'linear-gradient(135deg, rgba(34, 197, 94, 0.15), rgba(34, 197, 94, 0.05))',
-                                                    border: '1px solid rgba(34, 197, 94, 0.3)',
-                                                    padding: '18px',
-                                                    flexDirection: 'row',
-                                                    justifyContent: 'flex-start',
-                                                    gap: '14px',
-                                                    alignItems: 'center'
-                                                }}>
-                                                    <div style={{
-                                                        fontSize: '28px',
-                                                        width: '48px',
-                                                        height: '48px',
-                                                        background: 'rgba(34, 197, 94, 0.2)',
-                                                        borderRadius: '12px',
-                                                        display: 'flex',
-                                                        alignItems: 'center',
-                                                        justifyContent: 'center'
-                                                    }}>✅</div>
-                                                    <div>
-                                                        <div style={{ fontSize: '28px', fontWeight: 800, color: '#22c55e' }}>
-                                                            {epgTestResults.working.length}
-                                                        </div>
-                                                        <div style={{ color: 'rgba(255,255,255,0.5)', fontSize: '12px' }}>{t('epg', 'workingLabel')}</div>
-                                                    </div>
-                                                </div>
-                                                <div className="setting-item epg-grid-item" style={{
-                                                    background: 'linear-gradient(135deg, rgba(239, 68, 68, 0.15), rgba(239, 68, 68, 0.05))',
-                                                    border: '1px solid rgba(239, 68, 68, 0.3)',
-                                                    padding: '18px',
-                                                    flexDirection: 'row',
-                                                    justifyContent: 'flex-start',
-                                                    gap: '14px',
-                                                    alignItems: 'center'
-                                                }}>
-                                                    <div style={{
-                                                        fontSize: '28px',
-                                                        width: '48px',
-                                                        height: '48px',
-                                                        background: 'rgba(239, 68, 68, 0.2)',
-                                                        borderRadius: '12px',
-                                                        display: 'flex',
-                                                        alignItems: 'center',
-                                                        justifyContent: 'center'
-                                                    }}>❌</div>
-                                                    <div>
-                                                        <div style={{ fontSize: '28px', fontWeight: 800, color: '#ef4444' }}>
-                                                            {epgTestResults.notWorking.length}
-                                                        </div>
-                                                        <div style={{ color: 'rgba(255,255,255,0.5)', fontSize: '12px' }}>{t('epg', 'noDataLabel')}</div>
-                                                    </div>
-                                                </div>
-                                            </div>
-
-                                            {/* Country Counts Grid */}
-                                            <div style={{ display: 'grid', gridTemplateColumns: 'repeat(4, 1fr)', gap: '10px', marginBottom: '20px' }}>
-                                                {[
-                                                    { code: 'BR', flag: '🇧🇷', label: 'Brasil', color: '#22c55e' },
-                                                    { code: 'ARG', flag: '🇦🇷', label: 'Argentina', color: '#60a5fa' },
-                                                    { code: 'US', flag: '🇺🇸', label: 'EUA', color: '#f59e0b' },
-                                                    { code: 'PT', flag: '🇵🇹', label: 'Portugal', color: '#ef4444' }
-                                                ].map(({ code, flag, label, color }) => {
-                                                    const count = [...epgTestResults.working, ...epgTestResults.notWorking]
-                                                        .filter(c => c.country === code).length;
-                                                    return (
-                                                        <div key={code} className="setting-item epg-country-card" style={{
-                                                            padding: '16px 12px',
-                                                            background: `rgba(${color === '#22c55e' ? '34, 197, 94' : color === '#60a5fa' ? '96, 165, 250' : color === '#f59e0b' ? '245, 158, 11' : '239, 68, 68'}, 0.08)`,
-                                                            border: `1px solid rgba(${color === '#22c55e' ? '34, 197, 94' : color === '#60a5fa' ? '96, 165, 250' : color === '#f59e0b' ? '245, 158, 11' : '239, 68, 68'}, 0.2)`,
-                                                            justifyContent: 'center'
-                                                        }}
-                                                            onClick={() => setEpgCountryFilter(code as EpgCountryFilter)}
-                                                        >
-                                                            <div style={{ textAlign: 'center' }}>
-                                                                <div style={{ fontSize: '24px', marginBottom: '4px' }}>{flag}</div>
-                                                                <div style={{ fontSize: '20px', fontWeight: 800, color }}>
-                                                                    {count}
-                                                                </div>
-                                                                <div style={{ color: 'rgba(255,255,255,0.5)', fontSize: '11px' }}>
-                                                                    {label}
-                                                                </div>
-                                                            </div>
-                                                        </div>
-                                                    );
-                                                })}
-                                            </div>
-
-                                            {/* Filters - Reorganized */}
-                                            <div className="setting-item" style={{ flexDirection: 'column', alignItems: 'stretch', gap: '14px', marginBottom: '16px' }}>
-                                                {/* Row 1: Status Filters + Search */}
-                                                <div style={{ display: 'grid', gridTemplateColumns: 'auto 1fr', gap: '12px', alignItems: 'center' }}>
-                                                    <div style={{ display: 'flex', gap: '6px' }}>
-                                                        {(['all', 'working', 'notWorking'] as const).map((filter) => (
-                                                            <button
-                                                                key={filter}
-                                                                className="epg-filter-btn"
-                                                                onClick={() => { setEpgResultsFilter(filter); setEpgCurrentPage(1); }}
-                                                                style={{
-                                                                    padding: '8px 14px',
-                                                                    borderRadius: '8px',
-                                                                    border: epgResultsFilter === filter ? '2px solid #818cf8' : '1px solid rgba(255,255,255,0.1)',
-                                                                    cursor: 'pointer',
-                                                                    fontSize: '12px',
-                                                                    fontWeight: 600,
-                                                                    background: epgResultsFilter === filter ? 'rgba(99, 102, 241, 0.2)' : 'rgba(255,255,255,0.05)',
-                                                                    color: epgResultsFilter === filter ? '#a5b4fc' : 'rgba(255,255,255,0.6)'
-                                                                }}
-                                                            >
-                                                                {filter === 'all' ? '📺 Todos' : filter === 'working' ? '✅' : '❌'}
-                                                            </button>
-                                                        ))}
-                                                    </div>
-                                                    <input
-                                                        type="text"
-                                                        placeholder="🔍 Buscar canal..."
-                                                        value={epgSearchTerm}
-                                                        onChange={(e) => { setEpgSearchTerm(e.target.value); setEpgCurrentPage(1); }}
-                                                        className="setting-select"
-                                                        style={{ padding: '10px 14px', fontSize: '13px' }}
-                                                    />
-                                                </div>
-
-                                                {/* Row 2: Country Filters */}
-                                                <div style={{ display: 'flex', gap: '6px', flexWrap: 'wrap' }}>
-                                                    {(['all', 'BR', 'ARG', 'US', 'PT'] as const).map((country) => {
-                                                        const icons: Record<string, string> = { all: '🌍', BR: '🇧🇷', ARG: '🇦🇷', US: '🇺🇸', PT: '🇵🇹' };
-                                                        const labels: Record<string, string> = { all: 'Todos', BR: 'BR', ARG: 'ARG', US: 'US', PT: 'PT' };
-                                                        return (
-                                                            <button
-                                                                key={country}
-                                                                className="epg-filter-btn"
-                                                                onClick={() => { setEpgCountryFilter(country); setEpgCurrentPage(1); }}
-                                                                style={{
-                                                                    padding: '6px 12px',
-                                                                    borderRadius: '8px',
-                                                                    border: epgCountryFilter === country ? '2px solid #06b6d4' : '1px solid rgba(255,255,255,0.1)',
-                                                                    cursor: 'pointer',
-                                                                    fontSize: '11px',
-                                                                    fontWeight: 600,
-                                                                    background: epgCountryFilter === country ? 'rgba(6, 182, 212, 0.2)' : 'rgba(255,255,255,0.05)',
-                                                                    color: epgCountryFilter === country ? '#67e8f9' : 'rgba(255,255,255,0.6)'
-                                                                }}
-                                                            >
-                                                                {icons[country]} {labels[country]}
-                                                            </button>
-                                                        );
-                                                    })}
-                                                </div>
-                                            </div>
-
-                                            {/* Channel List with Pagination */}
-                                            {(() => {
-                                                const filteredItems = [
-                                                    ...epgTestResults.working.map((w): EpgWorkingItem => ({ ...w, type: 'working' })),
-                                                    ...epgTestResults.notWorking.map((n): EpgNotWorkingItem => ({ ...n, type: 'notWorking' }))
-                                                ]
-                                                    .filter(item =>
-                                                        (epgResultsFilter === 'all' || item.type === epgResultsFilter) &&
-                                                        (epgCountryFilter === 'all' || item.country === epgCountryFilter) &&
-                                                        (epgSearchTerm === '' || item.channel.toLowerCase().includes(epgSearchTerm.toLowerCase()))
-                                                    );
-
-                                                const totalPages = Math.ceil(filteredItems.length / EPG_ITEMS_PER_PAGE);
-                                                const currentPage = Math.min(epgCurrentPage, totalPages || 1);
-                                                const startIndex = (currentPage - 1) * EPG_ITEMS_PER_PAGE;
-                                                const endIndex = startIndex + EPG_ITEMS_PER_PAGE;
-                                                const displayedItems = filteredItems.slice(startIndex, endIndex);
-
-                                                return (
-                                                    <>
-                                                        {/* Counter Header with Pagination Info */}
-                                                        <div style={{
-                                                            display: 'flex',
-                                                            justifyContent: 'space-between',
-                                                            alignItems: 'center',
-                                                            marginBottom: '10px',
-                                                            paddingBottom: '10px',
-                                                            borderBottom: '1px solid rgba(255,255,255,0.1)'
-                                                        }}>
-                                                            <span style={{ color: 'rgba(255,255,255,0.5)', fontSize: '12px' }}>
-                                                                📋 {t('epg', 'showing')} <strong style={{ color: 'white' }}>{startIndex + 1}-{Math.min(endIndex, filteredItems.length)}</strong> {t('epg', 'of')} <strong style={{ color: 'white' }}>{filteredItems.length}</strong> {t('epg', 'channelsCount')}
-                                                            </span>
-                                                            {totalPages > 1 && (
-                                                                <span style={{ color: '#06b6d4', fontSize: '11px', fontWeight: 600 }}>
-                                                                    📄 {currentPage} / {totalPages}
-                                                                </span>
-                                                            )}
-                                                        </div>
-
-                                                        {/* List */}
-                                                        <div style={{
-                                                            maxHeight: '350px',
-                                                            overflowY: 'auto',
-                                                            display: 'flex',
-                                                            flexDirection: 'column',
-                                                            gap: '8px',
-                                                            paddingRight: '8px'
-                                                        }}>
-                                                            {displayedItems.map((item, i) => (
-                                                                <div key={startIndex + i} className="setting-item epg-channel-item" style={{
-                                                                    padding: '14px 16px',
-                                                                    background: item.type === 'working'
-                                                                        ? 'linear-gradient(135deg, rgba(34, 197, 94, 0.08), rgba(34, 197, 94, 0.02))'
-                                                                        : 'linear-gradient(135deg, rgba(239, 68, 68, 0.08), rgba(239, 68, 68, 0.02))',
-                                                                    border: `1px solid ${item.type === 'working' ? 'rgba(34, 197, 94, 0.15)' : 'rgba(239, 68, 68, 0.15)'}`,
-                                                                    flexDirection: 'column',
-                                                                    alignItems: 'stretch',
-                                                                    gap: '8px'
-                                                                }}>
-                                                                    <div style={{ display: 'flex', justifyContent: 'space-between', alignItems: 'center' }}>
-                                                                        <span style={{ color: 'white', fontWeight: 600, fontSize: '14px' }}>{item.channel}</span>
-                                                                        <span style={{
-                                                                            color: item.type === 'working' ? '#22c55e' : '#ef4444',
-                                                                            fontSize: '11px',
-                                                                            fontWeight: 600,
-                                                                            background: item.type === 'working' ? 'rgba(34, 197, 94, 0.15)' : 'rgba(239, 68, 68, 0.15)',
-                                                                            padding: '4px 10px',
-                                                                            borderRadius: '6px'
-                                                                        }}>
-                                                                            {item.type === 'working' ? `${item.programCount} prog` : item.reason}
-                                                                        </span>
-                                                                    </div>
-                                                                    <div style={{ color: 'rgba(255,255,255,0.4)', fontSize: '11px' }}>
-                                                                        <span style={{ color: '#06b6d4' }}>{item.source}</span>
-                                                                        <span style={{ margin: '0 6px' }}>•</span>
-                                                                        <code style={{ color: '#a78bfa', background: 'rgba(167, 139, 250, 0.1)', padding: '2px 5px', borderRadius: '3px', fontSize: '10px' }}>{item.epgId}</code>
-                                                                    </div>
-                                                                </div>
-                                                            ))}
-                                                        </div>
-
-                                                        {/* Pagination Controls */}
-                                                        {totalPages > 1 && (
-                                                            <div style={{
-                                                                display: 'flex',
-                                                                justifyContent: 'center',
-                                                                alignItems: 'center',
-                                                                gap: '12px',
-                                                                marginTop: '16px',
-                                                                paddingTop: '12px',
-                                                                borderTop: '1px solid rgba(255,255,255,0.1)'
-                                                            }}>
-                                                                <button
-                                                                    className="epg-page-btn"
-                                                                    onClick={() => setEpgCurrentPage(1)}
-                                                                    disabled={currentPage === 1}
-                                                                    style={{
-                                                                        padding: '8px 12px',
-                                                                        borderRadius: '8px',
-                                                                        border: 'none',
-                                                                        background: currentPage === 1 ? 'rgba(255,255,255,0.05)' : 'rgba(6, 182, 212, 0.2)',
-                                                                        color: currentPage === 1 ? 'rgba(255,255,255,0.3)' : '#67e8f9',
-                                                                        cursor: currentPage === 1 ? 'not-allowed' : 'pointer',
-                                                                        fontWeight: 600,
-                                                                        fontSize: '12px',
-                                                                        transition: 'all 0.2s ease'
-                                                                    }}
-                                                                >⏮️</button>
-                                                                <button
-                                                                    className="epg-page-btn"
-                                                                    onClick={() => setEpgCurrentPage(p => Math.max(1, p - 1))}
-                                                                    disabled={currentPage === 1}
-                                                                    style={{
-                                                                        padding: '8px 16px',
-                                                                        borderRadius: '8px',
-                                                                        border: 'none',
-                                                                        background: currentPage === 1 ? 'rgba(255,255,255,0.05)' : 'rgba(6, 182, 212, 0.2)',
-                                                                        color: currentPage === 1 ? 'rgba(255,255,255,0.3)' : '#67e8f9',
-                                                                        cursor: currentPage === 1 ? 'not-allowed' : 'pointer',
-                                                                        fontWeight: 600,
-                                                                        fontSize: '12px',
-                                                                        transition: 'all 0.2s ease'
-                                                                    }}
-                                                                >◀ {t('epg', 'previous')}</button>
-                                                                <span className="epg-page-indicator" style={{
-                                                                    color: 'white',
-                                                                    fontWeight: 700,
-                                                                    fontSize: '14px',
-                                                                    padding: '8px 16px',
-                                                                    background: 'rgba(99, 102, 241, 0.2)',
-                                                                    borderRadius: '8px'
-                                                                }}>
-                                                                    {currentPage} / {totalPages}
-                                                                </span>
-                                                                <button
-                                                                    className="epg-page-btn"
-                                                                    onClick={() => setEpgCurrentPage(p => Math.min(totalPages, p + 1))}
-                                                                    disabled={currentPage === totalPages}
-                                                                    style={{
-                                                                        padding: '8px 16px',
-                                                                        borderRadius: '8px',
-                                                                        border: 'none',
-                                                                        background: currentPage === totalPages ? 'rgba(255,255,255,0.05)' : 'rgba(6, 182, 212, 0.2)',
-                                                                        color: currentPage === totalPages ? 'rgba(255,255,255,0.3)' : '#67e8f9',
-                                                                        cursor: currentPage === totalPages ? 'not-allowed' : 'pointer',
-                                                                        fontWeight: 600,
-                                                                        fontSize: '12px',
-                                                                        transition: 'all 0.2s ease'
-                                                                    }}
-                                                                >{t('epg', 'next')} ▶</button>
-                                                                <button
-                                                                    className="epg-page-btn"
-                                                                    onClick={() => setEpgCurrentPage(totalPages)}
-                                                                    disabled={currentPage === totalPages}
-                                                                    style={{
-                                                                        padding: '8px 12px',
-                                                                        borderRadius: '8px',
-                                                                        border: 'none',
-                                                                        background: currentPage === totalPages ? 'rgba(255,255,255,0.05)' : 'rgba(6, 182, 212, 0.2)',
-                                                                        color: currentPage === totalPages ? 'rgba(255,255,255,0.3)' : '#67e8f9',
-                                                                        cursor: currentPage === totalPages ? 'not-allowed' : 'pointer',
-                                                                        fontWeight: 600,
-                                                                        fontSize: '12px',
-                                                                        transition: 'all 0.2s ease'
-                                                                    }}
-                                                                >⏭️</button>
-                                                            </div>
-                                                        )}
-                                                    </>
-                                                );
-                                            })()}
-                                        </div>
-                                    )}
-                                </div>
-                            </div>
+                            <EpgSection
+                                epgResultsFilter={epgResultsFilter}
+                                setEpgResultsFilter={setEpgResultsFilter}
+                                epgCountryFilter={epgCountryFilter}
+                                setEpgCountryFilter={setEpgCountryFilter}
+                                epgSearchTerm={epgSearchTerm}
+                                setEpgSearchTerm={setEpgSearchTerm}
+                                epgCurrentPage={epgCurrentPage}
+                                setEpgCurrentPage={setEpgCurrentPage}
+                            />
                         )}
 
                         {/* Statistics Section */}
-                        {activeSection === 'stats' && (
-                            <div className="section-card">
-                                <div className="section-header">
-                                    <div className="section-icon" style={{ background: 'linear-gradient(135deg, #8b5cf6, #7c3aed)' }}>📊</div>
-                                    <div>
-                                        <h2>{t('stats', 'title')}</h2>
-                                        <p>{t('stats', 'description')}</p>
-                                    </div>
-                                </div>
-
-                                <div className="settings-group">
-                                    {/* Main Stats Cards */}
-                                    <div style={{
-                                        display: 'grid',
-                                        gridTemplateColumns: 'repeat(2, 1fr)',
-                                        gap: '16px',
-                                        marginBottom: '24px'
-                                    }}>
-                                        {/* Total Watch Time */}
-                                        <div style={{
-                                            background: 'linear-gradient(135deg, rgba(139, 92, 246, 0.2), rgba(124, 58, 237, 0.1))',
-                                            borderRadius: '16px',
-                                            padding: '24px',
-                                            border: '1px solid rgba(139, 92, 246, 0.3)'
-                                        }}>
-                                            <div style={{ fontSize: '32px', marginBottom: '8px' }}>⏱️</div>
-                                            <div style={{
-                                                fontSize: '28px',
-                                                fontWeight: 700,
-                                                color: '#a78bfa',
-                                                marginBottom: '4px'
-                                            }}>
-                                                {formatWatchTime(usageStats?.totalWatchTimeThisMonth || 0)}
-                                            </div>
-                                            <div style={{ color: 'rgba(255,255,255,0.6)', fontSize: '13px' }}>
-                                                {t('stats', 'thisMonth')}
-                                            </div>
-                                        </div>
-
-                                        {/* Watch Streak */}
-                                        <div style={{
-                                            background: 'linear-gradient(135deg, rgba(251, 146, 60, 0.2), rgba(234, 88, 12, 0.1))',
-                                            borderRadius: '16px',
-                                            padding: '24px',
-                                            border: '1px solid rgba(251, 146, 60, 0.3)'
-                                        }}>
-                                            <div style={{ fontSize: '32px', marginBottom: '8px' }}>🔥</div>
-                                            <div style={{
-                                                fontSize: '28px',
-                                                fontWeight: 700,
-                                                color: '#fb923c',
-                                                marginBottom: '4px'
-                                            }}>
-                                                {usageStats?.watchStreak || 0} {t('stats', 'days')}
-                                            </div>
-                                            <div style={{ color: 'rgba(255,255,255,0.6)', fontSize: '13px' }}>
-                                                {t('stats', 'currentStreak')}
-                                            </div>
-                                        </div>
-                                    </div>
-
-                                    {/* Content Breakdown */}
-                                    <div style={{
-                                        background: 'rgba(255,255,255,0.03)',
-                                        borderRadius: '16px',
-                                        padding: '20px',
-                                        border: '1px solid rgba(255,255,255,0.05)',
-                                        marginBottom: '20px'
-                                    }}>
-                                        <h3 style={{ color: 'white', fontSize: '14px', fontWeight: 600, marginBottom: '16px' }}>
-                                            {t('stats', 'timeByType')}
-                                        </h3>
-                                        <div style={{ display: 'flex', gap: '12px' }}>
-                                            <div style={{ flex: 1, textAlign: 'center' }}>
-                                                <div style={{ fontSize: '24px', marginBottom: '4px' }}>🎬</div>
-                                                <div style={{ color: '#3b82f6', fontWeight: 600 }}>
-                                                    {formatWatchTime(usageStats?.contentBreakdown?.movies || 0)}
-                                                </div>
-                                                <div style={{ color: 'rgba(255,255,255,0.5)', fontSize: '11px' }}>{t('stats', 'movies')}</div>
-                                            </div>
-                                            <div style={{ flex: 1, textAlign: 'center' }}>
-                                                <div style={{ fontSize: '24px', marginBottom: '4px' }}>📺</div>
-                                                <div style={{ color: '#10b981', fontWeight: 600 }}>
-                                                    {formatWatchTime(usageStats?.contentBreakdown?.series || 0)}
-                                                </div>
-                                                <div style={{ color: 'rgba(255,255,255,0.5)', fontSize: '11px' }}>{t('stats', 'series')}</div>
-                                            </div>
-                                            <div style={{ flex: 1, textAlign: 'center' }}>
-                                                <div style={{ fontSize: '24px', marginBottom: '4px' }}>📡</div>
-                                                <div style={{ color: '#f59e0b', fontWeight: 600 }}>
-                                                    {formatWatchTime(usageStats?.contentBreakdown?.live || 0)}
-                                                </div>
-                                                <div style={{ color: 'rgba(255,255,255,0.5)', fontSize: '11px' }}>{t('stats', 'liveTV')}</div>
-                                            </div>
-                                        </div>
-                                    </div>
-
-                                    {/* Weekly Chart */}
-                                    <div style={{
-                                        background: 'rgba(255,255,255,0.03)',
-                                        borderRadius: '16px',
-                                        padding: '20px',
-                                        border: '1px solid rgba(255,255,255,0.05)'
-                                    }}>
-                                        <h3 style={{ color: 'white', fontSize: '14px', fontWeight: 600, marginBottom: '16px' }}>
-                                            {t('stats', 'last7Days')}
-                                        </h3>
-                                        <div style={{ display: 'flex', alignItems: 'flex-end', gap: '8px', height: '80px' }}>
-                                            {weeklyStats.map((day, i) => {
-                                                const height = Math.max(8, (day.totalSeconds / getMaxWeeklySeconds()) * 70);
-                                                const isToday = day.date === new Date().toISOString().split('T')[0];
-                                                return (
-                                                    <div key={i} style={{ flex: 1, textAlign: 'center' }}>
-                                                        <div
-                                                            style={{
-                                                                height: `${height}px`,
-                                                                background: isToday
-                                                                    ? 'linear-gradient(180deg, #8b5cf6, #6d28d9)'
-                                                                    : 'linear-gradient(180deg, rgba(139, 92, 246, 0.5), rgba(139, 92, 246, 0.2))',
-                                                                borderRadius: '4px',
-                                                                margin: '0 auto',
-                                                                width: '100%',
-                                                                maxWidth: '24px',
-                                                                transition: 'height 0.3s ease'
-                                                            }}
-                                                            title={`${formatWatchTime(day.totalSeconds)}`}
-                                                        />
-                                                        <div style={{
-                                                            color: isToday ? '#a78bfa' : 'rgba(255,255,255,0.4)',
-                                                            fontSize: '11px',
-                                                            marginTop: '6px',
-                                                            fontWeight: isToday ? 600 : 400
-                                                        }}>
-                                                            {getWeekDayLabel(day.date)}
-                                                        </div>
-                                                    </div>
-                                                );
-                                            })}
-                                        </div>
-                                    </div>
-
-                                    {/* Total All Time */}
-                                    <div style={{
-                                        marginTop: '16px',
-                                        padding: '16px',
-                                        background: 'rgba(255,255,255,0.02)',
-                                        borderRadius: '12px',
-                                        textAlign: 'center'
-                                    }}>
-                                        <span style={{ color: 'rgba(255,255,255,0.5)', fontSize: '13px' }}>
-                                            {t('stats', 'totalAccumulated')}: {' '}
-                                        </span>
-                                        <span style={{ color: 'white', fontWeight: 600 }}>
-                                            {formatWatchTime(usageStats?.totalWatchTimeSeconds || 0)}
-                                        </span>
-                                        {usageStats?.longestStreak && usageStats.longestStreak > 0 && (
-                                            <span style={{ color: 'rgba(255,255,255,0.5)', fontSize: '13px', marginLeft: '16px' }}>
-                                                • {t('stats', 'longestStreak')}: {usageStats.longestStreak} {t('stats', 'days')}
-                                            </span>
-                                        )}
-                                    </div>
-                                </div>
-                            </div>
-                        )}
+                        {activeSection === 'stats' && <StatsSection />}
 
                         {/* Parental Control Section */}
-                        {activeSection === 'parental' && (
-                            <div className="section-card">
-                                <div className="section-header">
-                                    <div className="section-icon" style={{ background: 'linear-gradient(135deg, #ef4444, #dc2626)' }}>👨‍👩‍👧</div>
-                                    <div>
-                                        <h2>{t('parental', 'title')}</h2>
-                                        <p>{t('parental', 'description')}</p>
-                                    </div>
-                                </div>
-
-                                <div className="settings-group">
-                                    <div className="setting-item">
-                                        <div className="setting-info">
-                                            <label>{t('parental', 'enable')}</label>
-                                            <p>{t('parental', 'enableDesc')}</p>
-                                        </div>
-                                        <label className="toggle-switch">
-                                            <input
-                                                type="checkbox"
-                                                checked={parentalConfig.enabled}
-                                                onChange={(e) => handleParentalConfigChange('enabled', e.target.checked)}
-                                            />
-                                            <span className="toggle-slider"></span>
-                                        </label>
-                                        {saveAnimation === 'parental_enabled' && <span className="save-indicator">{t('settings', 'saved')}</span>}
-                                    </div>
-
-                                    <div className="setting-item">
-                                        <div className="setting-info">
-                                            <label>{t('parental', 'maxRating')}</label>
-                                            <p>{t('parental', 'maxRatingDesc')}</p>
-                                        </div>
-                                        <select
-                                            className="setting-select"
-                                            value={parentalConfig.maxRating}
-                                            onChange={(e) => handleParentalConfigChange('maxRating', e.target.value as ParentalConfig['maxRating'])}
-                                            disabled={!parentalConfig.enabled}
-                                        >
-                                            <option value="L">{t('parental', 'free')}</option>
-                                            <option value="10">10 {t('parental', 'years')}</option>
-                                            <option value="12">12 {t('parental', 'years')}</option>
-                                            <option value="14">14 {t('parental', 'years')}</option>
-                                            <option value="16">16 {t('parental', 'years')}</option>
-                                            <option value="18">18 {t('parental', 'years')}</option>
-                                        </select>
-                                        {saveAnimation === 'parental_maxRating' && <span className="save-indicator">{t('settings', 'saved')}</span>}
-                                    </div>
-
-                                    <div className="setting-item">
-                                        <div className="setting-info">
-                                            <label>{t('parental', 'pin')}</label>
-                                            <p>{parentalService.hasPin() ? t('parental', 'pinConfigured') : t('parental', 'pinDefine')}</p>
-                                        </div>
-                                        <button
-                                            className="setting-btn"
-                                            onClick={() => {
-                                                resetPinModal();
-                                                setShowPinModal(true);
-                                            }}
-                                            disabled={!parentalConfig.enabled}
-                                            style={{
-                                                padding: '10px 20px',
-                                                background: parentalConfig.enabled ? 'rgba(239, 68, 68, 0.2)' : 'rgba(100, 100, 100, 0.2)',
-                                                border: `1px solid ${parentalConfig.enabled ? 'rgba(239, 68, 68, 0.4)' : 'rgba(100, 100, 100, 0.4)'}`,
-                                                borderRadius: '10px',
-                                                color: parentalConfig.enabled ? '#ef4444' : '#666',
-                                                cursor: parentalConfig.enabled ? 'pointer' : 'not-allowed',
-                                                fontWeight: 600,
-                                                transition: 'all 0.2s'
-                                            }}
-                                        >
-                                            🔑 {parentalService.hasPin() ? t('parental', 'changePin') : t('parental', 'setPin')} PIN
-                                        </button>
-                                        {saveAnimation === 'parental_pin' && <span className="save-indicator">{t('settings', 'saved')}</span>}
-                                    </div>
-
-                                    <div className="setting-item">
-                                        <div className="setting-info">
-                                            <label>{t('parental', 'blockAdult')}</label>
-                                            <p>{t('parental', 'blockAdultDesc')}</p>
-                                        </div>
-                                        <label className="toggle-switch">
-                                            <input
-                                                type="checkbox"
-                                                checked={parentalConfig.blockAdultCategories}
-                                                onChange={(e) => handleParentalConfigChange('blockAdultCategories', e.target.checked)}
-                                                disabled={!parentalConfig.enabled}
-                                            />
-                                            <span className="toggle-slider"></span>
-                                        </label>
-                                        {saveAnimation === 'parental_blockAdultCategories' && <span className="save-indicator">{t('settings', 'saved')}</span>}
-                                    </div>
-
-                                    <div className="setting-item">
-                                        <div className="setting-info">
-                                            <label>{t('parental', 'filterTMDB')}</label>
-                                            <p>{t('parental', 'filterTMDBDesc')}</p>
-                                        </div>
-                                        <label className="toggle-switch">
-                                            <input
-                                                type="checkbox"
-                                                checked={parentalConfig.filterByTMDB}
-                                                onChange={(e) => handleParentalConfigChange('filterByTMDB', e.target.checked)}
-                                                disabled={!parentalConfig.enabled}
-                                            />
-                                            <span className="toggle-slider"></span>
-                                        </label>
-                                        {saveAnimation === 'parental_filterByTMDB' && <span className="save-indicator">{t('settings', 'saved')}</span>}
-                                    </div>
-                                </div>
-                            </div>
-                        )}
+                        {activeSection === 'parental' && <ParentalSection />}
 
                         {/* About Section */}
-                        {activeSection === 'about' && (
-                            <div className="section-card">
-                                <div className="section-header">
-                                    <div className="section-icon" style={{ background: 'linear-gradient(135deg, #f59e0b, #d97706)' }}>ℹ️</div>
-                                    <div>
-                                        <h2>{t('about', 'title')}</h2>
-                                        <p>{t('about', 'description')}</p>
-                                    </div>
-                                </div>
-
-                                <div className="about-content">
-                                    <div className="app-logo">
-                                        <svg width="80" height="80" viewBox="0 0 100 100" xmlns="http://www.w3.org/2000/svg">
-                                            <defs>
-                                                <linearGradient id="logoGrad" x1="0%" y1="0%" x2="100%" y2="100%">
-                                                    <stop offset="0%" stopColor="#a855f7" />
-                                                    <stop offset="100%" stopColor="#ec4899" />
-                                                </linearGradient>
-                                            </defs>
-                                            <path d="M 10,10 L 10,90 L 90,50 Z" fill="none" stroke="url(#logoGrad)" strokeWidth="6" strokeLinejoin="round" />
-                                            <rect x="35" y="35" width="6" height="30" fill="url(#logoGrad)" rx="3" />
-                                            <rect x="45" y="25" width="6" height="50" fill="url(#logoGrad)" rx="3" />
-                                            <rect x="55" y="40" width="6" height="20" fill="url(#logoGrad)" rx="3" />
-                                        </svg>
-                                    </div>
-                                    <h3 className="app-name">NeoStream</h3>
-                                    <p className="app-version">{t('about', 'version')} {__APP_VERSION__}</p>
-                                    <p className="app-description">
-                                        {t('about', 'appDescription')}
-                                    </p>
-                                    <div className="about-links">
-                                        <button
-                                            onClick={() => setShowTermsModal(true)}
-                                            className="about-link"
-                                        >
-                                            📄 {t('about', 'termsOfUse')}
-                                        </button>
-                                        <button
-                                            onClick={() => setShowPrivacyModal(true)}
-                                            className="about-link"
-                                        >
-                                            🔒 {t('about', 'privacyPolicy')}
-                                        </button>
-                                        <a href="mailto:suporte@neostream.app" className="about-link">💬 {t('about', 'support')}</a>
-                                    </div>
-                                </div>
-                            </div>
-                        )}
+                        {activeSection === 'about' && <AboutSection />}
                     </div>
                 </div>
-
-                {/* PIN Modal */}
-                {showPinModal && (
-                    <div
-                        className="pin-modal-overlay"
-                        style={{
-                            position: 'fixed',
-                            inset: 0,
-                            background: 'rgba(0, 0, 0, 0.9)',
-                            backdropFilter: 'blur(12px)',
-                            display: 'flex',
-                            alignItems: 'center',
-                            justifyContent: 'center',
-                            zIndex: 1000,
-                            animation: 'pinFadeIn 0.3s ease'
-                        }}
-                        onClick={(e) => {
-                            if (e.target === e.currentTarget) {
-                                setShowPinModal(false);
-                                resetPinModal();
-                            }
-                        }}
-                    >
-                        <style>{`
-                            @keyframes pinFadeIn {
-                                from { opacity: 0; }
-                                to { opacity: 1; }
-                            }
-                            @keyframes pinSlideIn {
-                                from { opacity: 0; transform: translateY(-30px) scale(0.95); }
-                                to { opacity: 1; transform: translateY(0) scale(1); }
-                            }
-                            @keyframes pinBounce {
-                                0%, 100% { transform: scale(1); }
-                                50% { transform: scale(1.15); }
-                            }
-                            @keyframes pinDigitPop {
-                                0% { transform: scale(1); }
-                                50% { transform: scale(1.1); }
-                                100% { transform: scale(1); }
-                            }
-                            @keyframes pinShake {
-                                0%, 100% { transform: translateX(0); }
-                                20% { transform: translateX(-10px); }
-                                40% { transform: translateX(10px); }
-                                60% { transform: translateX(-10px); }
-                                80% { transform: translateX(10px); }
-                            }
-                            .pin-digit-filled {
-                                animation: pinDigitPop 0.2s ease !important;
-                            }
-                            .pin-error-shake {
-                                animation: pinShake 0.5s ease !important;
-                            }
-                        `}</style>
-                        <div style={{
-                            background: 'linear-gradient(145deg, #1e293b 0%, #0f172a 100%)',
-                            borderRadius: '28px',
-                            padding: '48px 40px',
-                            maxWidth: '420px',
-                            width: '90%',
-                            border: '1px solid rgba(239, 68, 68, 0.3)',
-                            boxShadow: '0 30px 100px rgba(0, 0, 0, 0.6), 0 0 40px rgba(239, 68, 68, 0.1)',
-                            animation: 'pinSlideIn 0.4s cubic-bezier(0.34, 1.56, 0.64, 1)'
-                        }}>
-                            {/* Header */}
-                            <div style={{ textAlign: 'center', marginBottom: '36px' }}>
-                                <span style={{
-                                    fontSize: '56px',
-                                    display: 'block',
-                                    marginBottom: '16px',
-                                    animation: 'pinBounce 0.6s ease'
-                                }}>{pinMode === 'verify' ? '🔓' : '🔐'}</span>
-                                <h2 style={{
-                                    color: 'white',
-                                    fontSize: '26px',
-                                    fontWeight: 700,
-                                    margin: '0 0 10px 0',
-                                    background: 'linear-gradient(135deg, #fff 0%, #f87171 100%)',
-                                    WebkitBackgroundClip: 'text',
-                                    WebkitTextFillColor: 'transparent',
-                                    backgroundClip: 'text'
-                                }}>
-                                    {pinMode === 'verify'
-                                        ? t('parental', 'verifyPin')
-                                        : pinStep === 'enter'
-                                            ? t('parental', 'setPin') + ' PIN'
-                                            : t('parental', 'confirmPin')}
-                                </h2>
-                                <p style={{ color: '#9ca3af', fontSize: '15px', margin: 0 }}>
-                                    {pinMode === 'verify'
-                                        ? t('parental', 'verifyPin')
-                                        : pinStep === 'enter'
-                                            ? t('parental', 'enterPin')
-                                            : t('parental', 'confirmPin')}
-                                </p>
-                            </div>
-
-                            {/* PIN Input Container */}
-                            <div
-                                style={{
-                                    position: 'relative',
-                                    display: 'flex',
-                                    justifyContent: 'center',
-                                    gap: '14px',
-                                    marginBottom: '28px'
-                                }}
-                                onClick={() => {
-                                    const input = document.getElementById('pin-hidden-input');
-                                    if (input) input.focus();
-                                }}
-                            >
-                                {/* Hidden Input that captures keyboard */}
-                                <input
-                                    id="pin-hidden-input"
-                                    type="tel"
-                                    inputMode="numeric"
-                                    pattern="[0-9]*"
-                                    maxLength={4}
-                                    autoFocus
-                                    autoComplete="off"
-                                    value={pinStep === 'enter' ? pinInput.join('') : pinConfirm.join('')}
-                                    onChange={(e) => {
-                                        const value = e.target.value.replace(/\D/g, '').slice(0, 4);
-                                        const digits = value.split('').concat(['', '', '', '']).slice(0, 4);
-                                        if (pinStep === 'enter') {
-                                            setPinInput(digits);
-                                        } else {
-                                            setPinConfirm(digits);
-                                        }
-                                        setPinError('');
-                                    }}
-                                    style={{
-                                        position: 'absolute',
-                                        top: 0,
-                                        left: 0,
-                                        width: '100%',
-                                        height: '100%',
-                                        opacity: 0,
-                                        cursor: 'pointer',
-                                        zIndex: 10
-                                    }}
-                                />
-
-                                {/* Visual PIN Digits */}
-                                {(pinStep === 'enter' ? pinInput : pinConfirm).map((digit, index) => (
-                                    <div
-                                        key={index}
-                                        className={digit ? 'pin-digit-filled' : ''}
-                                        style={{
-                                            width: '64px',
-                                            height: '76px',
-                                            background: digit
-                                                ? 'linear-gradient(135deg, rgba(239, 68, 68, 0.2) 0%, rgba(220, 38, 38, 0.15) 100%)'
-                                                : 'rgba(255, 255, 255, 0.05)',
-                                            border: `2px solid ${digit ? '#ef4444' : 'rgba(255, 255, 255, 0.15)'}`,
-                                            borderRadius: '16px',
-                                            display: 'flex',
-                                            alignItems: 'center',
-                                            justifyContent: 'center',
-                                            fontSize: '36px',
-                                            color: '#ef4444',
-                                            transition: 'all 0.2s ease',
-                                            boxShadow: digit ? '0 4px 20px rgba(239, 68, 68, 0.3)' : 'none',
-                                            cursor: 'pointer'
-                                        }}
-                                    >
-                                        {digit ? '●' : ''}
-                                    </div>
-                                ))}
-                            </div>
-
-                            {/* Error Message */}
-                            {pinError && (
-                                <div
-                                    className="pin-error-shake"
-                                    style={{
-                                        display: 'flex',
-                                        alignItems: 'center',
-                                        justifyContent: 'center',
-                                        gap: '8px',
-                                        color: '#f87171',
-                                        fontSize: '14px',
-                                        fontWeight: 500,
-                                        marginBottom: '20px',
-                                        padding: '12px',
-                                        background: 'rgba(239, 68, 68, 0.1)',
-                                        borderRadius: '10px',
-                                        border: '1px solid rgba(239, 68, 68, 0.2)'
-                                    }}
-                                >
-                                    ⚠️ {pinError}
-                                </div>
-                            )}
-
-                            {/* Buttons */}
-                            <div style={{
-                                display: 'flex',
-                                gap: '14px',
-                                marginTop: '28px'
-                            }}>
-                                <button
-                                    type="button"
-                                    onClick={() => {
-                                        setShowPinModal(false);
-                                        resetPinModal();
-                                        // If we're enabling parental control and user cancels, disable it
-                                        if (!parentalService.hasPin()) {
-                                            handleParentalConfigChange('enabled', false);
-                                        }
-                                    }}
-                                    style={{
-                                        flex: 1,
-                                        padding: '16px 24px',
-                                        borderRadius: '14px',
-                                        fontSize: '15px',
-                                        fontWeight: 600,
-                                        cursor: 'pointer',
-                                        background: 'rgba(255, 255, 255, 0.08)',
-                                        color: '#9ca3af',
-                                        border: '1px solid rgba(255, 255, 255, 0.1)',
-                                        transition: 'all 0.2s ease'
-                                    }}
-                                    onMouseEnter={(e) => {
-                                        e.currentTarget.style.background = 'rgba(255, 255, 255, 0.12)';
-                                        e.currentTarget.style.color = 'white';
-                                    }}
-                                    onMouseLeave={(e) => {
-                                        e.currentTarget.style.background = 'rgba(255, 255, 255, 0.08)';
-                                        e.currentTarget.style.color = '#9ca3af';
-                                    }}
-                                >
-                                    Cancelar
-                                </button>
-                                <button
-                                    type="button"
-                                    onClick={() => handlePinSubmit()}
-                                    style={{
-                                        flex: 1,
-                                        padding: '16px 24px',
-                                        borderRadius: '14px',
-                                        fontSize: '15px',
-                                        fontWeight: 600,
-                                        cursor: 'pointer',
-                                        background: 'linear-gradient(135deg, #ef4444 0%, #dc2626 100%)',
-                                        color: 'white',
-                                        border: 'none',
-                                        transition: 'all 0.2s ease',
-                                        boxShadow: '0 4px 20px rgba(239, 68, 68, 0.4)'
-                                    }}
-                                    onMouseEnter={(e) => {
-                                        e.currentTarget.style.transform = 'translateY(-2px)';
-                                        e.currentTarget.style.boxShadow = '0 6px 25px rgba(239, 68, 68, 0.5)';
-                                    }}
-                                    onMouseLeave={(e) => {
-                                        e.currentTarget.style.transform = 'translateY(0)';
-                                        e.currentTarget.style.boxShadow = '0 4px 20px rgba(239, 68, 68, 0.4)';
-                                    }}
-                                >
-                                    {pinMode === 'verify'
-                                        ? '🔓 Desbloquear'
-                                        : pinStep === 'enter'
-                                            ? 'Continuar →'
-                                            : '✓ Confirmar'}
-                                </button>
-                            </div>
-                        </div>
-                    </div>
-                )}
-
-                {/* Terms of Use Modal */}
-                {showTermsModal && (
-                    <div
-                        style={{
-                            position: 'fixed',
-                            inset: 0,
-                            background: 'rgba(0, 0, 0, 0.9)',
-                            backdropFilter: 'blur(12px)',
-                            display: 'flex',
-                            alignItems: 'center',
-                            justifyContent: 'center',
-                            zIndex: 1000,
-                            padding: '20px',
-                            animation: 'modalFadeIn 0.3s ease'
-                        }}
-                        onClick={(e) => {
-                            if (e.target === e.currentTarget) setShowTermsModal(false);
-                        }}
-                    >
-                        <style>{`
-                            @keyframes modalFadeIn {
-                                from { opacity: 0; }
-                                to { opacity: 1; }
-                            }
-                            @keyframes modalSlideIn {
-                                from { opacity: 0; transform: translateY(-40px) scale(0.95); }
-                                to { opacity: 1; transform: translateY(0) scale(1); }
-                            }
-                            @keyframes iconPulse {
-                                0%, 100% { transform: scale(1); }
-                                50% { transform: scale(1.1); }
-                            }
-                            @keyframes fadeInUp {
-                                from { opacity: 0; transform: translateY(20px); }
-                                to { opacity: 1; transform: translateY(0); }
-                            }
-                        `}</style>
-                        <div style={{
-                            background: 'linear-gradient(145deg, #1e293b 0%, #0f172a 100%)',
-                            borderRadius: '24px',
-                            padding: '32px',
-                            maxWidth: '700px',
-                            width: '100%',
-                            maxHeight: '80vh',
-                            overflow: 'auto',
-                            border: '1px solid rgba(168, 85, 247, 0.3)',
-                            boxShadow: '0 25px 80px rgba(0, 0, 0, 0.5), 0 0 40px rgba(168, 85, 247, 0.1)',
-                            animation: 'modalSlideIn 0.4s cubic-bezier(0.34, 1.56, 0.64, 1)'
-                        }}>
-                            <div style={{ display: 'flex', justifyContent: 'space-between', alignItems: 'center', marginBottom: '24px' }}>
-                                <h2 style={{
-                                    color: 'white',
-                                    fontSize: '24px',
-                                    fontWeight: 700,
-                                    margin: 0,
-                                    display: 'flex',
-                                    alignItems: 'center',
-                                    gap: '12px'
-                                }}>
-                                    <span style={{ fontSize: '32px', animation: 'iconPulse 2s ease infinite' }}>📄</span>
-                                    <span style={{
-                                        background: 'linear-gradient(135deg, #a855f7 0%, #ec4899 100%)',
-                                        WebkitBackgroundClip: 'text',
-                                        WebkitTextFillColor: 'transparent',
-                                        backgroundClip: 'text'
-                                    }}>Termos de Uso</span>
-                                </h2>
-                                <button
-                                    onClick={() => setShowTermsModal(false)}
-                                    style={{
-                                        background: 'rgba(168, 85, 247, 0.15)',
-                                        border: '1px solid rgba(168, 85, 247, 0.3)',
-                                        borderRadius: '12px',
-                                        padding: '10px 16px',
-                                        color: '#a855f7',
-                                        cursor: 'pointer',
-                                        fontSize: '14px',
-                                        fontWeight: 600,
-                                        transition: 'all 0.2s ease'
-                                    }}
-                                    onMouseEnter={(e) => {
-                                        e.currentTarget.style.background = 'rgba(168, 85, 247, 0.25)';
-                                        e.currentTarget.style.transform = 'scale(1.05)';
-                                    }}
-                                    onMouseLeave={(e) => {
-                                        e.currentTarget.style.background = 'rgba(168, 85, 247, 0.15)';
-                                        e.currentTarget.style.transform = 'scale(1)';
-                                    }}
-                                >
-                                    ✕ Fechar
-                                </button>
-                            </div>
-                            <div style={{ color: '#9ca3af', fontSize: '14px', lineHeight: 1.8, animation: 'fadeInUp 0.5s ease 0.2s both' }}>
-                                <p style={{ marginBottom: '16px', padding: '12px 16px', background: 'rgba(168, 85, 247, 0.1)', borderRadius: '12px', border: '1px solid rgba(168, 85, 247, 0.2)' }}>
-                                    <strong style={{ color: '#a855f7' }}>📅 Última atualização:</strong> <span style={{ color: 'white' }}>09 de Dezembro de 2025</span>
-                                </p>
-
-                                <h3 style={{ color: '#a855f7', fontSize: '16px', marginTop: '24px', marginBottom: '12px' }}>1. Aceitação dos Termos</h3>
-                                <p>Ao utilizar o NeoStream, você concorda com estes Termos de Uso. O aplicativo é destinado exclusivamente para uso pessoal e não comercial.</p>
-
-                                <h3 style={{ color: '#a855f7', fontSize: '16px', marginTop: '24px', marginBottom: '12px' }}>2. Descrição do Serviço</h3>
-                                <p>O NeoStream é um player de mídia que permite visualizar conteúdo IPTV através de listas M3U fornecidas pelo usuário. Não fornecemos, hospedamos ou distribuímos qualquer conteúdo de mídia.</p>
-
-                                <h3 style={{ color: '#a855f7', fontSize: '16px', marginTop: '24px', marginBottom: '12px' }}>3. Responsabilidade do Usuário</h3>
-                                <p>Você é responsável por:</p>
-                                <ul style={{ marginLeft: '20px', marginTop: '8px' }}>
-                                    <li>Garantir que possui direitos legais sobre o conteúdo acessado</li>
-                                    <li>Cumprir as leis de direitos autorais do seu país</li>
-                                    <li>Manter suas credenciais de acesso seguras</li>
-                                    <li>Usar o aplicativo de forma ética e responsável</li>
-                                </ul>
-
-                                <h3 style={{ color: '#a855f7', fontSize: '16px', marginTop: '24px', marginBottom: '12px' }}>4. Uso do Controle Parental</h3>
-                                <p>O recurso de controle parental é fornecido como ferramenta auxiliar. Os pais/responsáveis devem supervisionar o uso do aplicativo por menores.</p>
-
-                                <h3 style={{ color: '#a855f7', fontSize: '16px', marginTop: '24px', marginBottom: '12px' }}>5. Limitação de Responsabilidade</h3>
-                                <p>O NeoStream é fornecido "como está", sem garantias. Não nos responsabilizamos por:</p>
-                                <ul style={{ marginLeft: '20px', marginTop: '8px' }}>
-                                    <li>Conteúdo de terceiros acessado através do aplicativo</li>
-                                    <li>Interrupções no serviço de streaming</li>
-                                    <li>Perdas de dados ou problemas técnicos</li>
-                                </ul>
-
-                                <h3 style={{ color: '#a855f7', fontSize: '16px', marginTop: '24px', marginBottom: '12px' }}>6. Modificações</h3>
-                                <p>Reservamo-nos o direito de modificar estes termos a qualquer momento. Alterações significativas serão comunicadas através do aplicativo.</p>
-                            </div>
-                        </div>
-                    </div>
-                )}
-
-                {/* Privacy Policy Modal */}
-                {showPrivacyModal && (
-                    <div
-                        style={{
-                            position: 'fixed',
-                            inset: 0,
-                            background: 'rgba(0, 0, 0, 0.9)',
-                            backdropFilter: 'blur(12px)',
-                            display: 'flex',
-                            alignItems: 'center',
-                            justifyContent: 'center',
-                            zIndex: 1000,
-                            padding: '20px',
-                            animation: 'modalFadeIn 0.3s ease'
-                        }}
-                        onClick={(e) => {
-                            if (e.target === e.currentTarget) setShowPrivacyModal(false);
-                        }}
-                    >
-                        <style>{`
-                            @keyframes modalFadeIn {
-                                from { opacity: 0; }
-                                to { opacity: 1; }
-                            }
-                            @keyframes modalSlideIn {
-                                from { opacity: 0; transform: translateY(-40px) scale(0.95); }
-                                to { opacity: 1; transform: translateY(0) scale(1); }
-                            }
-                            @keyframes iconPulse {
-                                0%, 100% { transform: scale(1); }
-                                50% { transform: scale(1.1); }
-                            }
-                            @keyframes fadeInUp {
-                                from { opacity: 0; transform: translateY(20px); }
-                                to { opacity: 1; transform: translateY(0); }
-                            }
-                        `}</style>
-                        <div style={{
-                            background: 'linear-gradient(145deg, #1e293b 0%, #0f172a 100%)',
-                            borderRadius: '24px',
-                            padding: '32px',
-                            maxWidth: '700px',
-                            width: '100%',
-                            maxHeight: '80vh',
-                            overflow: 'auto',
-                            border: '1px solid rgba(16, 185, 129, 0.3)',
-                            boxShadow: '0 25px 80px rgba(0, 0, 0, 0.5), 0 0 40px rgba(16, 185, 129, 0.1)',
-                            animation: 'modalSlideIn 0.4s cubic-bezier(0.34, 1.56, 0.64, 1)'
-                        }}>
-                            <div style={{ display: 'flex', justifyContent: 'space-between', alignItems: 'center', marginBottom: '24px' }}>
-                                <h2 style={{
-                                    color: 'white',
-                                    fontSize: '24px',
-                                    fontWeight: 700,
-                                    margin: 0,
-                                    display: 'flex',
-                                    alignItems: 'center',
-                                    gap: '12px'
-                                }}>
-                                    <span style={{ fontSize: '32px', animation: 'iconPulse 2s ease infinite' }}>🔒</span>
-                                    <span style={{
-                                        background: 'linear-gradient(135deg, #10b981 0%, #059669 100%)',
-                                        WebkitBackgroundClip: 'text',
-                                        WebkitTextFillColor: 'transparent',
-                                        backgroundClip: 'text'
-                                    }}>Política de Privacidade</span>
-                                </h2>
-                                <button
-                                    onClick={() => setShowPrivacyModal(false)}
-                                    style={{
-                                        background: 'rgba(16, 185, 129, 0.15)',
-                                        border: '1px solid rgba(16, 185, 129, 0.3)',
-                                        borderRadius: '12px',
-                                        padding: '10px 16px',
-                                        color: '#10b981',
-                                        cursor: 'pointer',
-                                        fontSize: '14px',
-                                        fontWeight: 600,
-                                        transition: 'all 0.2s ease'
-                                    }}
-                                    onMouseEnter={(e) => {
-                                        e.currentTarget.style.background = 'rgba(16, 185, 129, 0.25)';
-                                        e.currentTarget.style.transform = 'scale(1.05)';
-                                    }}
-                                    onMouseLeave={(e) => {
-                                        e.currentTarget.style.background = 'rgba(16, 185, 129, 0.15)';
-                                        e.currentTarget.style.transform = 'scale(1)';
-                                    }}
-                                >
-                                    ✕ Fechar
-                                </button>
-                            </div>
-                            <div style={{ color: '#9ca3af', fontSize: '14px', lineHeight: 1.8, animation: 'fadeInUp 0.5s ease 0.2s both' }}>
-                                <p style={{ marginBottom: '16px', padding: '12px 16px', background: 'rgba(16, 185, 129, 0.1)', borderRadius: '12px', border: '1px solid rgba(16, 185, 129, 0.2)' }}>
-                                    <strong style={{ color: '#10b981' }}>📅 Última atualização:</strong> <span style={{ color: 'white' }}>09 de Dezembro de 2025</span>
-                                </p>
-
-                                <h3 style={{ color: '#10b981', fontSize: '16px', marginTop: '24px', marginBottom: '12px' }}>1. Dados que Coletamos</h3>
-                                <p>O NeoStream coleta e armazena <strong style={{ color: 'white' }}>localmente no seu dispositivo</strong>:</p>
-                                <ul style={{ marginLeft: '20px', marginTop: '8px' }}>
-                                    <li>Credenciais de conexão IPTV (criptografadas)</li>
-                                    <li>Preferências de configuração</li>
-                                    <li>Histórico de reprodução e progresso</li>
-                                    <li>Configurações de controle parental</li>
-                                    <li>Cache de classificações de conteúdo</li>
-                                </ul>
-
-                                <h3 style={{ color: '#10b981', fontSize: '16px', marginTop: '24px', marginBottom: '12px' }}>2. Armazenamento Local</h3>
-                                <p>Todos os dados são armazenados <strong style={{ color: 'white' }}>exclusivamente no seu dispositivo</strong> usando:</p>
-                                <ul style={{ marginLeft: '20px', marginTop: '8px' }}>
-                                    <li>LocalStorage para configurações</li>
-                                    <li>IndexedDB para cache de conteúdo</li>
-                                    <li>SessionStorage para dados temporários</li>
-                                </ul>
-
-                                <h3 style={{ color: '#10b981', fontSize: '16px', marginTop: '24px', marginBottom: '12px' }}>3. Comunicações Externas</h3>
-                                <p>O aplicativo faz conexões externas apenas para:</p>
-                                <ul style={{ marginLeft: '20px', marginTop: '8px' }}>
-                                    <li>Seu servidor IPTV (fornecido por você)</li>
-                                    <li>API do TMDB para metadados e classificações</li>
-                                    <li>Verificação de atualizações (opcional)</li>
-                                </ul>
-
-                                <h3 style={{ color: '#10b981', fontSize: '16px', marginTop: '24px', marginBottom: '12px' }}>4. Não Coletamos</h3>
-                                <p style={{ color: '#10b981' }}>❌ NÃO coletamos, transmitimos ou vendemos:</p>
-                                <ul style={{ marginLeft: '20px', marginTop: '8px' }}>
-                                    <li>Informações pessoais identificáveis</li>
-                                    <li>Dados de localização</li>
-                                    <li>Histórico de navegação</li>
-                                    <li>Informações de contatos</li>
-                                    <li>Dados analíticos ou telemetria</li>
-                                </ul>
-
-                                <h3 style={{ color: '#10b981', fontSize: '16px', marginTop: '24px', marginBottom: '12px' }}>5. Seus Direitos</h3>
-                                <p>Você pode a qualquer momento:</p>
-                                <ul style={{ marginLeft: '20px', marginTop: '8px' }}>
-                                    <li>Limpar todos os dados locais nas configurações</li>
-                                    <li>Desinstalar o aplicativo para remover todos os dados</li>
-                                    <li>Exportar seus dados (histórico, favoritos)</li>
-                                </ul>
-
-                                <h3 style={{ color: '#10b981', fontSize: '16px', marginTop: '24px', marginBottom: '12px' }}>6. Contato</h3>
-                                <p>Para questões sobre privacidade, entre em contato através do suporte do aplicativo.</p>
-                            </div>
-                        </div>
-                    </div>
-                )}
             </div >
         </>
     );
@@ -2115,7 +131,7 @@ const settingsStyles = `
 .settings-backdrop {
     position: fixed;
     inset: 0;
-    background: 
+    background:
         radial-gradient(ellipse at 30% 30%, rgba(168, 85, 247, 0.1) 0%, transparent 50%),
         radial-gradient(ellipse at 70% 70%, rgba(59, 130, 246, 0.1) 0%, transparent 50%);
     pointer-events: none;
